Validators 30/4/20 22:35

+++ comoEnCasa/comoEnCasa.COMMON/Entidades/ventas.cs
@@ -9,6 +9,6 @@ namespace comoEnCasa.COMMON.Entidades
         public int IdVenta { get; set; }
         public DateTime FechaVenta { get; set; }
         public decimal ValorTotalVenta { get; set; }
-        public short DescuentoVenta { get; set; }
+        public byte DescuentoVenta { get; set; }
     }
 }
